Fix Karma datatype

+++ EliteAPI/Models/DTOs/Outgoing/Player.cs
@@ -13,7 +13,7 @@ public class PlayerDataDto
     public long LastLogin { get; set; }
     public long LastLogout { get; set; }
 
-    public int Karma { get; set; }
+    public long Karma { get; set; }
     public double NetworkExp { get; set; }
    
     public int RewardHighScore { get; set; }
